atc/db: Reset resource config IDs on save
When a pipeline is updated, the resources and their config is updated in
the database. However, the `resource_config_id` and
`resource_config_scope_id` fields are not updated and so the build
inputs algorithm considers "stale" versions (versions associated with an
old resource config) as valid.

By setting these values back to NULL, we force LIDAR to FindOrCreate an
existing resource_config and resource_config_scope (create new if config
has changed) whilst also preventing the stale versions from being
considered. As LIDAR is triggered immediately after SavePipeline and
LIDAR does the FindOrCreate before trying to schedule any builds, this
means that any existing versions for resources that haven't changed are
quickly "reattached" and become valid again.

Fixes #7468

Co-Authored-By: Richard Inskip <richard.inskip@depop.com>
Signed-off-by: Andy Paine <andy.paine@engineerbetter.com>