atc/db: Only decrypt config if blob has changed
Don't both decrypting the config to compare the `source` fields unless
the blobs are different. Encryption should be stable and it should be
much quicker to just compare the strings and only decrypt when we have
to. Comparing the strings themselves is not sufficient as some parts of
the encrypted object are allowed to change without versions being reset

Signed-off-by: Andy Paine <andy.paine@engineerbetter.com>