Reword streaming volume message
Be explicit that the source is a worker in messages (in web and fly
stream) and omit the destination since the `selected-worker` event shows
that and this would just be duplicate info.

I've left the DestWorker on the actual event since it could be useful as
part of debugging or an audit trail. Omitting it in output makes sense
as all the events around it give context but by leaving the DestWorker
in, this means you don't need to "piece together" what happened before
the `streaming-volume` event and can just look at it in isolation

Signed-off-by: Andy Paine <andy.paine@engineerbetter.com>