atc: Avoid duplicating volume streams
When atc.EnableCacheStreamedVolumes is true, streamed volumes are marked
as a resource cache on a worker. If multiple steps that require the same
volume are assigned to the same worker in parallel, each would
(previously) stream the volume.

To avoid adding excessive load to the source worker with the volume, we
can instead just create a lock based on volume+worker and other steps
can wait for the volume to be streamed and marked as a cache. This
avoids duplicating work for both source and destination worker.

Co-Authored-By: Richard Inskip richard.inskip@depop.com
Signed-off-by: Andy Paine <andy.paine@engineerbetter.com>